feat: extract blob storage container management from BlobUrlService
